[pkg/stanza] Cleanup assignkeys operator files
